% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.prob.R
\name{estimate.blrm}
\alias{estimate.blrm}
\title{Bayesian Logistic Regression Model (BLRM) for modeling the posterior toxicity probability}
\usage{
estimate.blrm(y, n, d, plot = TRUE)
}
\arguments{
\item{y}{Integer vector: Number of accumulated dose-limiting toxicities (DLTs) at each dose.}

\item{n}{Integer vector: Number of accumulated patients treated at each dose.}

\item{d}{Numeric vector: vector of dose level}

\item{plot}{Boolean: show plot of the estimated posterior probability}

\item{target}{Numeric: Target toxicity rate.}
}
\value{
a dataframe containing Posterior Mean and 95\% Credible Interval for each dose
}
\description{
The function \code{estimate.blrm} estimates Posterior Toxicity Probabilities using observed dose-limiting toxicities (DLT) y and number of enrolled patients n.
The function was adapted based on the work of Ruitao Lin.
}
\examples{
dose <- c(1, 2, 3, 4, 5);y <- c(1, 2, 3, 4, 5);n <- c(3, 6, 9, 12, 15); estimate.blrm(y,n,dose)
library(rstanarm)  # uncomment this for standalone use
library(ggplot2)
}
