% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.rules.R
\name{get.rules}
\alias{get.rules}
\title{Escalation and De-escalation threshold Generator}
\usage{
get.rules(target, ncohort, cohortsize = 3, design = 7, cutoff.eli = 0.95)
}
\arguments{
\item{target}{Numeric: The target dose-limiting toxicity (DLT) rate; default=0.3.}

\item{ncohort}{Numeric: The total number of cohorts; scalar.}

\item{cohortsize}{Integer: The number of subjects per cohort; scalar, default=3.}

\item{design}{Integer: The interval designs of dose finding to be used; scalar. \cr
Default values for parameters in designs are used. \cr
\itemize{
\item 1 - Central Composite (CCD) Design
\item 2 - Modified Toxicity Probability Interval (mTPI) Design
\item 3 - Bayesian Optimal Interva (BOIN) Design
\item 4 - Keyboard or mTPI2 deisgn
\item 5 - Uniformly Most Powerful Bayesian Interval (UMPBI) Design
}}

\item{cutoff.eli}{Numeric: Cutoff probability of early termination,
avoiding the overly toxic dose for safety monitoring; scalar, default=0.95}
}
\value{
A table/matrix of escalation, de-escalation, and elimination rules for each cohort.
}
\description{
The function \code{get.rules} generates escalation and de-escalation rules
for several interval designs including the CCD, mTPI,
BOIN, Keyboard and UMPBI designs.
The function was adapted based on the work of Ruitao Lin
}
\examples{
get.rules(target=0.3,ncohort=10, cohortsize=3, design=2, cutoff.eli=0.95) # this should match Fig 2 of mTPI original paper
}
