% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pava.transform.R
\name{pava}
\alias{pava}
\title{Pool Adjacent Violator Algorithm (PAVA)}
\usage{
pava(x, wt = rep(1, length(x)))
}
\arguments{
\item{x}{Numeric vector: Input values.}

\item{wt}{Numeric vector: Weights (default is uniform weight).}
}
\value{
Numeric vector: Isotonically adjusted values.
}
\description{
an efficient isotonic regression/smoothing method that enforces a monotonic constraint on a sequence of values, e.g., toxicity probabilities for dose-response modeling. The function was adapted based on the work of Ruitao Lin
}
\examples{
# toxicity_rates <- c(0.05, 0.10, 0.15, 0.14, 0.20) \cr
# pava(toxicity_rates) \cr
#[1] 0.05 0.10 0.15 0.15 0.20 \cr
}
