% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.mtd.R
\name{select.mtd}
\alias{select.mtd}
\title{Select Maximum Tolerated Dose (MTD) in a Dose-Finding Trial}
\usage{
select.mtd(target, y, n, type = 2, cutoff.eli = 0.95)
}
\arguments{
\item{target}{Numeric: Target toxicity rate.}

\item{y}{Integer vector: Number of accumulated dose-limiting toxicities (DLTs) at each dose.}

\item{n}{Integer vector: Number of accumulated patients treated at each dose.}

\item{type}{Integer: The method for estimating posterior toxicity probabilities. \cr
\itemize{
\item 1 - PAVA regression with a weakly informative conjugate prior with Beta(0.05,0.05)
\item 2 - Bayesian Logistic Regression Model (BLRM) with a weakly informative normal prior
}}

\item{cutoff.eli}{Numeric: Threshold probability for eliminating an overly toxic dose (default = 0.95).}
}
\value{
Integer: Selected MTD dose level (or 99 if no dose is safe).
}
\description{
The function \code{select.mtd} estimates Posterior Toxicity Probabilities and determines the MTD based on toxicity data.
The function was adapted based on the work of Ruitao Lin.
}
