% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.summary.R
\name{summary.simulation}
\alias{summary.simulation}
\title{Operating Characteristics for Bayesian Dose-Finding Designs}
\usage{
\method{summary}{simulation}(
  target,
  p.true,
  ncohort,
  cohortsize,
  startdose = 1,
  design = 5,
  type = 2,
  cutoff.eli = 0.95,
  ntrial = 1000
)
}
\arguments{
\item{target}{Numeric: Target toxicity rate.}

\item{p.true}{Numeric vector: True toxicity probability for each dose.}

\item{ncohort}{Integer: Number of patient cohorts.}

\item{cohortsize}{Integer: Number of patients per cohort.}

\item{startdose}{Integer: Initial dose level.}

\item{design}{Integer: Estimation choice (1=PAVA, 2=BLRM).}

\item{cutoff.eli}{Numeric: Cutoff probability for dose elimination (default = 0.95).}

\item{ntrial}{Integer: Number of simulated trials.}
}
\value{
A data frame summarizing trial results; A list of summary
}
\description{
Simulates clinical trials using interval-based dose-finding method. The function was adapted based on the work of Ruitao Lin
}
\examples{
p.true<-c(0.08,0.10,0.20,0.30,0.45,0.60); \br
summary.simulation(target=0.3,p.true=p.true,ncohort=12,cohortsize=3,type =2,design=5,cutoff.eli=0.95)
}
